Add faster version to fit a linear model.
For now, it was decided to leave the version of fitting the linear model through lm() since the documentation mentions that speedlm.fit() can have some numerical instabilities. However, the automated tests showed no disruption. Therefore, it must be tested in other circumstances and see the direct changes.
The benchmarks showed an overall 5x increase. For instance, for the full set of expression assessment with the full set of dorothea it was reduced from 35 minutes to 7 minutes.